fix userrepo wrong userid type and make BCryptPasswordEncoder as static

+++ src/main/java/com/aicourse/repo/UserRepo.java
@@ -5,8 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepo extends JpaRepository<Users, Integer> {
+public interface UserRepo extends JpaRepository<Users, Long> {
 
     Users findByUsername(String username);
-
 }
